new review seeder and purger
review seeder:
this creates random reviews for each food under each user to fill the database with information to test with

review purger:
this removes all test reviews from the review seeder. the seeder marks reviews with a test field that the purger detects and deletes based on its value.